src/agents/order/requirements.txt
langgraph==1.0.2

azure-ai-agentserver-langgraph==1.0.0b3

pytest==8.4.2
azure-identity==1.25.0
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.8.1